Ensure that volume mounts survive temporary disconnects to the remote.
When using Telepresence for docker compose, it's not uncommon to stop
and start containers that are engaged in such a way that they provide
volume mounts from the remote host. A `stop` ends the engagement and a
`start` reengages. Any mounts made for remote volumes must then find the
volume again using the same IP and port, even if the remote host has
been temporarily unavailable.

This commit ensures that info about the remote volume is kept until the
volume is deleted so that a reconnection attempt is made. A failure to
reconnect will make the driver forget the remote.

Signed-off-by: Thomas Hallgren <thomas@tada.se>